requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
yt-dlp>=2024.1.0
zipstream-ng>=1.8.0
aiofiles>=24.1.0
